TagParser.cs was splitted. TokenReader.cs и Converter.cs were refactored.
