+++ cs/Markdown/Markdown.csproj
@@ -47,8 +47,10 @@
     <Compile Include="Converter\Converter.cs" />
     <Compile Include="Converter\MarkdownToHtmlConverter.cs" />
     <Compile Include="Md.cs" />
+    <Compile Include="Parser\InlineTagParser.cs" />
     <Compile Include="Parser\ITagParser.cs" />
-    <Compile Include="Parser\TagParser.cs" />
+    <Compile Include="Parser\LineTagParser.cs" />
+    <Compile Include="Parser\UnorderedListTagParser.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tags\HtmlTagStorage.cs" />
